requirements.txt
streamlit
streamlit_chat
sentence_transformers
pinecone
langchain_community